refactor: re-enable the wrong-import-order pylint check
The Google pylintrc disables it, but I can't find any reasoning in
their style guide. And I strongly prefer sane ordered import
sections. So I'm re-enabling it for now. If we hate the strictness,
we can revert.

While I'm at it, rename test/ to tests/ to avoid conflicting with
the standard library's test module.