feat!: switch to async http request methods
This changes the following API methods to be async:
- client.post
- client.extract
- transformer.list_polarity
- transformer.map_polarity

They all also take an optional client parameter, to specify an
httpx.AsyncClient.

This also drops support for python 3.7 for its lack of good async
testing support (and it's about to be EOL anyway).

Version has been bumped to 3.0.0, as this is a major breaking change.